+++ src/processing/steps/field-validation/list-item.js
@@ -1,35 +1,6 @@
 const valuePrep = require("../../../common/value-prep");
-const inputErrors = require("../../../common/input-errors");
 
-
-function validateListItem(rowInd, rowObj, colName, maxLengthNum, arrayName, fullResObj)
-{
-	var givenValue = rowObj[colName];
-	var correctType = valuePrep.checkString(givenValue);
-	var validationResult = -1;
-	
-	if (correctType === true && givenValue.length > 0 && givenValue.length <= maxLengthNum)
-	{
-		validationResult = addItem(givenValue, arrayName, fullResObj);
-	}
-	else if (correctType === true && givenValue.length > maxLengthNum)
-	{
-		inputErrors.setStringTooLong(fullResObj.error, colName, rowInd, maxLengthNum);
-	}
-	else if (correctType === true)
-	{
-		inputErrors.setStringEmpty(fullResObj.error, colName, rowInd);
-	}
-	else
-	{
-		inputErrors.setInvalidType(fullResObj.error, colName, rowInd, "string");
-	}
-	
-	return validationResult;
-}
-
-
-function addItem(itemString, arrName, fullRes)
+function addListItem(itemString, arrName, fullRes)
 {
 	var arrObj = fullRes.data[arrName];
 	var preparedItem = valuePrep.sanitizeString(itemString);
@@ -69,5 +40,5 @@ function addItem(itemString, arrName, fullRes)
 
 module.exports =
 {
-	validateItem: validateListItem
+	addItem: addListItem
 };