+++ src/processing/steps/field-validation/sub-location.js
@@ -1,39 +1,7 @@
 const valuePrep = require("../../../common/value-prep");
-const inputErrors = require("../../../common/input-errors");
 
 
-function validateSubLocation(rowInd, rowObj, colName, maxLengthNum, arrayName, parentLocationID, fullResObj, entryRequired)
-{
-	var givenValue = rowObj[colName];
-	var correctType = valuePrep.checkString(givenValue);
-	var validationResult = -1;
-	
-	if (correctType === true && givenValue.length > 0 && givenValue.length <= maxLengthNum)
-	{
-		validationResult = addLocation(givenValue, arrayName, parentLocationID, fullResObj);
-	}
-	else if (correctType === true && givenValue.length > maxLengthNum)
-	{
-		inputErrors.setStringTooLong(fullResObj.error, colName, rowInd, maxLengthNum);
-	}
-	else if (correctType === true && entryRequired === true)
-	{
-		inputErrors.setStringEmpty(fullResObj.error, colName, rowInd);
-	}
-	else if (correctType === true)
-	{
-		validationResult = addLocation("", arrayName, parentLocationID, fullResObj);
-	}
-	else
-	{
-		inputErrors.setInvalidType(fullResObj.error, colName, rowInd, "string");
-	}
-	
-	return validationResult;
-}
-
-
-function addLocation(locationString, arrName, targetParentID, fullRes)
+function addLocationItem(locationString, arrName, targetParentID, fullRes)
 {
 	var arrObj = fullRes.data[arrName];
 	var preparedItem = valuePrep.sanitizeString(locationString);
@@ -79,5 +47,5 @@ function addLocation(locationString, arrName, targetParentID, fullRes)
 
 module.exports =
 {
-	validateLocation: validateSubLocation
+	addLocation: addLocationItem
 };