+++ src/processing/steps/general-pass.js
@@ -20,7 +20,7 @@ function loopDataRows(origData, fullResultObj)
 	while (rowLoopIndex >= 0 && rowLoopIndex < origData.length && fullResultObj.canContinue === true)
 	{
 		currentRowObject = origData[rowLoopIndex];
-		currentTerritory = listItem.validateItem(rowLoopIndex, currentRowObject, "TERRITORY", valueLimits.territory, "territories", fullResultObj);
+		currentTerritory = -1;
 		currentCountry = -1;
 		currentState = -1;
 		currentCity = -1;
@@ -29,6 +29,7 @@ function loopDataRows(origData, fullResultObj)
 		currentOrderStatus = -1;
 		currentRowComplete = false;
 		
+		currentTerritory = handleTerritoryNormalization(rowLoopIndex, currentRowObject, fullResultObj);
 		currentCountry = handleCountryNormalization(rowLoopIndex, currentRowObject, currentTerritory, fullResultObj);
 		currentState = handleStateNormalization(rowLoopIndex, currentRowObject, currentCountry, fullResultObj);
 		currentCity = handleCityNormalization(rowLoopIndex, currentRowObject, currentState, fullResultObj);
@@ -52,13 +53,34 @@ function loopDataRows(origData, fullResultObj)
 }
 
 
+
+function handleTerritoryNormalization(rowIndex, rowObject, fullResult)
+{
+	var stringObject = stringValue.validateString(rowIndex, rowObject, "TERRITORY", valueLimits.territory, true, fullResult);
+	var handleRes = -1;
+	
+	if (stringObject.valid === true)
+	{
+		handleRes = listItem.addItem(stringObject.preparedText, "territories", fullResult);
+	}
+	
+	return handleRes;
+}
+
+
 function handleCountryNormalization(rowIndex, rowObject, territoryValue, fullResult)
 {
+	var stringObject = {};
 	var handleRes = -1;
 	
 	if (territoryValue > 0 && territoryValue <= fullResult.data.territories.length)
 	{
-		handleRes = countryItem.validateCountry(rowIndex, rowObject, territoryValue, fullResult);
+		stringObject = stringValue.validateString(rowIndex, rowObject, "COUNTRY", valueLimits.country, true, fullResult);
+	}
+	
+	if (stringObject.valid === true)
+	{
+		handleRes = countryItem.addCountry(stringObject.preparedText, fullResult.data.countries, territoryValue);
 	}
 	
 	return handleRes;
@@ -67,11 +89,17 @@ function handleCountryNormalization(rowIndex, rowObject, territoryValue, fullRes
 
 function handleStateNormalization(rowIndex, rowObject, countryValue, fullResult)
 {
+	var stringObject = {};
 	var handleRes = -1;
 	
 	if (countryValue > 0 && countryValue <= fullResult.data.countries.length)
 	{
-		handleRes = subLocation.validateLocation(rowIndex, rowObject, "STATE", valueLimits.state, "statesRegions", countryValue, fullResult, false);
+		stringObject = stringValue.validateString(rowIndex, rowObject, "STATE", valueLimits.state, false, fullResult);
+	}
+	
+	if (stringObject.valid === true)
+	{
+		handleRes = subLocation.addLocation(stringObject.preparedText, "statesRegions", countryValue, fullResult);
 	}
 	
 	return handleRes;
@@ -80,11 +108,17 @@ function handleStateNormalization(rowIndex, rowObject, countryValue, fullResult)
 
 function handleCityNormalization(rowIndex, rowObject, stateRegionValue, fullResult)
 {
+	var stringObject = {};
 	var handleRes = -1;
 	
 	if (stateRegionValue > 0 && stateRegionValue <= fullResult.data.statesRegions.length)
 	{
-		handleRes = subLocation.validateLocation(rowIndex, rowObject, "CITY", valueLimits.city, "cities", stateRegionValue, fullResult, true);
+		stringObject = stringValue.validateString(rowIndex, rowObject, "CITY", valueLimits.city, true, fullResult);
+	}
+	
+	if (stringObject.valid === true)
+	{
+		handleRes = subLocation.addLocation(stringObject.preparedText, "cities", stateRegionValue, fullResult);
 	}
 	
 	return handleRes;
@@ -93,23 +127,36 @@ function handleCityNormalization(rowIndex, rowObject, stateRegionValue, fullResu
 
 function handleDealSizeNormalization(rowIndex, rowObject, cityValue, fullResult)
 {
+	var stringObject = {};
 	var handleRes = -1;
 	
 	if (cityValue > 0 && cityValue <= fullResult.data.cities.length)
 	{
-		handleRes = listItem.validateItem(rowIndex, rowObject, "DEALSIZE", valueLimits.dealSize, "dealSizes", fullResult);
+		stringObject = stringValue.validateString(rowIndex, rowObject, "DEALSIZE", valueLimits.dealSize, true, fullResult);
+	}
+	
+	if (stringObject.valid === true)
+	{
+		handleRes = listItem.addItem(stringObject.preparedText, "dealSizes", fullResult);
 	}
 	
 	return handleRes;
 }
 
+
 function handleProductLineNormalization(rowIndex, rowObject, dealSizeValue, fullResult)
 {
+	var stringObject = {};
 	var handleRes = -1;
 	
 	if (dealSizeValue > 0 && dealSizeValue <= fullResult.data.dealSizes.length)
 	{
-		handleRes = listItem.validateItem(rowIndex, rowObject, "PRODUCTLINE", valueLimits.productLine, "productLines", fullResult);
+		stringObject = stringValue.validateString(rowIndex, rowObject, "PRODUCTLINE", valueLimits.productLine, true, fullResult);
+	}
+	
+	if (stringObject.valid === true)
+	{
+		handleRes = listItem.addItem(stringObject.preparedText, "productLines", fullResult);
 	}
 	
 	return handleRes;
@@ -118,11 +165,17 @@ function handleProductLineNormalization(rowIndex, rowObject, dealSizeValue, full
 
 function handleOrderStatusNormalization(rowIndex, rowObject, productLineValue, fullResult)
 {
+	var stringObject = {};
 	var handleRes = -1;
 	
 	if (productLineValue > 0 && productLineValue <= fullResult.data.productLines.length)
 	{
-		handleRes = listItem.validateItem(rowIndex, rowObject, "STATUS", valueLimits.orderStatus, "orderStatusModes", fullResult);
+		stringObject = stringValue.validateString(rowIndex, rowObject, "STATUS", valueLimits.orderStatus, true, fullResult);
+	}
+	
+	if (stringObject.valid === true)
+	{
+		handleRes = listItem.addItem(stringObject.preparedText, "orderStatusModes", fullResult);
 	}
 	
 	return handleRes;